Agent table with actions
